Contact now interface & class cluster. No more reflection.
