+++ api/src/main/java/de/unistuttgart/iaas/amyassist/amy/core/plugin/api/registry/IRegistry.java
@@ -13,7 +13,7 @@ public interface IRegistry<T> {
      * Get all entities of this registry
      * @return all entities
      */
-    List<T> getAll();
+    List<? extends T> getAll();
 
     /**
      * Get the entity with this exact id
